refactor: remove helmut.splunk.base
