+++ pytest_splunk_addon/helmut/splunk/__init__.py
@@ -13,12 +13,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""
-This module contains everything regarding a Splunk installation.
-
-@author: Nicklas Ansman-Giertz
-@contact: U{ngiertz@splunk.com<mailto:ngiertz@splunk.com>}
-@since: 2011-11-23
-"""
-
-__all__ = ["local", "ssh"]
